added OAuth code
I was able to add the OAuth code to implement in the program. when the user clicks on the login button, they are redirected to the google login page. At the moment the page pops up, but the user is not able to select their account, I think this is due to a link mismatch. Accounts are also not being saved to the DB, so will need to check that